Refactor the shard map parsing so we can pass it down into the API object.
Right now the type itself is defined in the cli package, which means we can't
use it without an import cycle.
